+++ web/modules/custom/hoeringsportal_dialogue/src/Helper/DialogueHelper.php
@@ -234,7 +234,7 @@ class DialogueHelper {
       if (!in_array('use_age_on_proposals', $config)) {
         $form['field_age_span']['#access'] = FALSE;
       }
-      
+
       if (!in_array('use_name_on_proposals', $config)) {
         $form['field_owner_name']['#access'] = FALSE;
       }