+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-587](https://github.com/itk-dev/deltag.aarhus.dk/pull/587)
+  Cleaned up translations
 * [PR-586](https://github.com/itk-dev/hoeringsportal/pull/586)
   * Added course content type
   * Added activity wrapper for public_events and courses